requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
huggingface_hub==0.14.1
langchain>=0.1.0
langchain-pinecone
pinecone-client>=3.0.0
flask
pypdf
python-dotenv
-e .